backend_django/pyproject.toml
[tool.poetry]
name = "backend-django"
version = "0.1.0"
description = ""
authors = ["shurxet"]
readme = "README.md"
packages = [{include = "backend_django"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.2"
djangorestframework = "3.12"
drf-spectacular = "^0.26.1"
django-cors-headers = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "4.3.0"
djoser = "^2.1.0"
psycopg2-binary = "^2.9.5"
phonenumbers = "^8.13.7"
django-phonenumber-field = "^7.0.2"
django-filter = "^22.1"
pillow = "^9.4.0"
pyjwt = "v1.7.1"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
